+++ src/app/list/ShoppingRecipeItem.tsx
@@ -3,6 +3,17 @@
 import { Button } from "~/components/ui/button";
 import { useShoppingListActions } from "../useShoppingListActions";
 import { Trash } from "lucide-react";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "~/components/ui/alert-dialog";
 
 export function ShoppingRecipeItem(props: { id: string; name: string }) {
   const { id, name } = props;
@@ -11,15 +22,34 @@ export function ShoppingRecipeItem(props: { id: string; name: string }) {
 
   return (
     <div key={id} className="flex items-center gap-2">
-      <Button
-        onClick={async () => {
-          await handleDeleteRecipe(Number(id));
-        }}
-        variant="destructive-outline"
-        size="sm"
-      >
-        <Trash />
-      </Button>
+      <AlertDialog>
+        <AlertDialogTrigger asChild>
+          <Button variant="destructive-outline" size="sm">
+            <Trash />
+          </Button>
+        </AlertDialogTrigger>
+        <AlertDialogContent>
+          <AlertDialogHeader>
+            <AlertDialogTitle>
+              Are you sure you want to delete?
+            </AlertDialogTitle>
+            <AlertDialogDescription>
+              This will remove all items for this recipe from your shopping
+              list.
+            </AlertDialogDescription>
+          </AlertDialogHeader>
+          <AlertDialogFooter>
+            <AlertDialogCancel>Cancel</AlertDialogCancel>
+            <AlertDialogAction
+              onClick={async () => {
+                await handleDeleteRecipe(Number(id));
+              }}
+            >
+              Delete
+            </AlertDialogAction>
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialog>
       <span className="break-words text-lg font-semibold">{name}</span>
     </div>
   );
